mcp-server/requirements.txt
fastapi>=0.111.3
uvicorn>=0.23.2
httpx>=0.25.1
python-dotenv>=1.0.0
pydantic>=2.8.0
